MPI Segmentation fault fixed
- Index of the compressed realization fixed
- Index of the `in_cone` check fixed

+++ src/libcal_mpi/include/cal/CAL_MPI_AtmSim.hpp
@@ -77,7 +77,7 @@ class mpi_atm_sim{
             double xstep = 100, double ystep = 100, double zstep = 100,
 
             /** Size of the simulation slices [elements]*/
-            uint64_t nelem_sim_max = 1000,
+            long nelem_sim_max = 1000,
 
             /**Verbosity [bool]*/
             int verbosity = 0, MPI_Comm comm = MPI_COMM_WORLD,
@@ -140,7 +140,7 @@ class mpi_atm_sim{
 
         double xstrideinv, ystrideinv, zstrideinv;
 
-        uint64_t nelem;
+        size_t nelem;
 
         bool cached = false;
 
@@ -150,7 +150,7 @@ class mpi_atm_sim{
         long nr;
 
         /**Size of the independent X-direction slices*/
-        uint64_t nelem_sim_max;
+        long nelem_sim_max;
 
         /**Kolmogorov correlation grid*/
         double rmin_kolmo, rmax_kolmo, rstep, rstep_inv;
@@ -180,7 +180,7 @@ class mpi_atm_sim{
         mpi_shmem_double * realization = NULL;
 
         /**Find the next range of compressed indices to simulate*/
-        void get_slice(uint64_t & ind_start, uint64_t & ind_stop);
+        void get_slice(long & ind_start, long & ind_stop);
 
         /**
         * Use the atmospheric parameters for volume element covariance
